Fix Yarn project (#370)
The `react/yarn3` project did not seem to be configured correctly to use
Plug'n'Play. Performing `yarn install` in the repo was still creating a
`node_modules` directory and there was no `.pnp.cjs` or
`.pnp.loader.mjs` file present.

This simply recreates the project from scratch and ensures it's using
PnP mode.

Also renamed the projects to just be `yarn` instead of `yarn3` as we
will probably upgrade Yarn in the future.