Add CORS to Node.js examples (#463)
This PR refactors the Connect for Node.js examples. The major changes
are:

- The examples in the `/node.js` directory have all been moved to the
top-level along with the other projects.
- Each Node example now contains a full CORS setup so that the project
can be run with a separate web interface using CORS.
- In addition, the examples also server the frontend interface via Node,
so they can be accessed from a web browser without the need for CORS.

Thanks to @akinorimizushima and his work on #444, which inspired this
CORS refactoring. The `fastify` example in #444 has been ported to this
PR.