Add an example using RTL with mocked transports to react/cra (#911)
This adds an example of component testing using React Testing Library to
the Create React App project.

In doing so, it refactors the components in the project so that a
`Transport` interface is passed as a prop. This way, users can easily
use `createRouterTransport` to create a mock transport and pass that to
the component in their test. This approach can then be suggested in docs
as a recommendation for easily testing components that interact with
Connect clients and transports.

Note that this also moves the actual test execution out of
`react-app-rewired` and into pure Jest. This is because we need to set
the `testEnvironment` to `@bufbuild/jest-environment-jsdom` because of
the issues with `jsdom`. However, neither CRA nor `react-app-rewired`
seem to allow you to override this value despite docs indicating it is
possible. It is also just more intuitive and easy to reason about to use
Jest directly.