Add bundle-size testing directory (#1365)
This adds a `bundle-size` directory which assists in testing Connect and
Protobuf-ES with various bundlers. It will enable us to test how changes
to those libraries affect bundle size and tree-shaking across an array
of popular bundlers in the ecosystem.

Note that this directory is not meant to show an example of any sort as
the others do. This is just for our own testing purposes.

---------

Co-authored-by: Timo Stamm <ts@timostamm.de>