Add additional testing with bundle sizes for TypeScript (#1444)
This expands the `bundle-size` directory to include testing and size
calculations for both JavaScript and TypeScript. Currently, the code
only calculates bundle size for JavaScript and generated JS files from
Protobuf-ES and it does not verify that the code generated by the
bundlers actually runs.

Now, each subdirectory for the bundlers has been expanded to generate JS
files as well as TS+DTS files. The bundlers bundle each as separate
entry points and then calculate the bundle size for TypeScript as well
as JavaScript. In addition, it runs the code generated by the bundlers
for both TS and JS and verifies that imports are correct and the code
runs.

This is helpful to test code that returns type information at runtime.
For example, it allows us to now verify that adding `PURE` annotations
to generated TypeScript for enums will throw errors at runtime.