Update dependabot config with new directories (#1488)
This updates the dependabot.yml config to add some directories that were
missing / incorrect.

Note:
- Since we have the `manage.mjs` script, it is not totally necessary to
have every directory in dependabot.yml, but keeping it in sync has two
benefits: 1. it cross-checks that our `manage.mjs` script is accurately
updating all dependencies in the repo (otherwise those dependabot PRs
will stay open). 2. For completeness sake since it feels odd to have
only some of our directories in that config.
- The `bundle-size` has also been added and each bundler is its own
group. Currently, the `manage.mjs` script does not handle workspaces,
which is how the `bundle-size` directory is structured, so we actually
_need_ that directory to be managed by dependabot. We could update
`manage.mjs` to handle the workspace setup in the future if need be.