Enhance manage script for npm workspaces (#1490)
This adds some additional functionality to the `manage.mjs` script to
handle npm workspaces. Currently, the `bundle-size` directory is
somewhat of an anomaly in the repo. It's main purpose is testing and is
structured using npm workspaces. This meant that the previous
incarnation of the manage script would not perform the necessary
package.json updates when `forceupdateknown` and `forceupdateall` were
called.

This modifies the script so that any package that contains workspaces
maintains the workspaces as part of the `PackageEnt` class and knows how
to recursively call the workspace update functions.

---------

Co-authored-by: Timo Stamm <ts@timostamm.de>