Fix Vite CJS build warning for Yarn and Vue (#1722)
Fixes #1721 

This fixes the warning about Vite's deprecated CJS build in the
`react/yarn-pnp` and `vue` projects by converting them to an ESM module.
Previously, they were both CJS, most likely due to a simple oversight.

With Vue, it also involved renaming the Cypress config to be a `.js`
file. Vue scaffolds Cypress config files to be a `.ts` file, which in
turn does not work with ESM modules. See [this
issue](https://github.com/cypress-io/cypress/issues/23552) for more
info.

For more info on the Vite CJS message:
https://vitejs.dev/guide/troubleshooting.html#vite-cjs-node-api-deprecated

---------

Signed-off-by: Steve Ayers <sayers@buf.build>