Update repo deps for August 2024 (#1787)
This updates repo deps for August 2024 for all deps except:
* `@bufbuild/protobuf`
* `@bufbuild/protoc-gen-es`
* `@bufbuild/protoplugin`

This is because the above deps have a released v2, but the currently
released `@connectrpc` packages are not yet updated to accept v2 as a
peer dep, so updating the above packages to their latest would result in
peer dependency errors.

Also, the Svelte project's update of eslint caused errors because of its
usage of `.eslintrc.cjs`. The instructions
[here](https://eslint.org/docs/latest/use/configure/migration-guide#migrate-your-config-file)
were followed to migrate that project's config.

---------

Signed-off-by: Steve Ayers <sayers@buf.build>