Update React Native example (#1791)
This is a combination of the efforts in #1706 and in #1708. It updates
the React Native example with latest deps as well as a better approach
to polyfills. We only realized recently that #1708 was a fork and wasn't
merged into main, so this remedies that. 😬

Thanks to @ryanlntn for the polyfill work in #1708.

---------

Signed-off-by: Steve Ayers <sayers@buf.build>