Fix React Native subpath imports (#2133)
Using the Metro bundler with React Native requires enabling package
exports ([docs](https://metrobundler.dev/docs/package-exports/)) via a
Metro config file. This adds said file and some necessary dependencies.

Note that we also spell this out in our Protobuf-ES
[docs](https://github.com/bufbuild/protobuf-es/blob/main/MANUAL.md#parcelmetro-fails-to-resolve-imports).

It seems odd to have to use the `react-native-svg-transformer` here. But
without it, we get asset errors. This combination here appears to be the
right incantation to get things to work. Taken from this
[issue](https://github.com/facebook/react-native/issues/38282).

---------

Signed-off-by: Steve Ayers <sayers@buf.build>